Project 1 First IOT Sensor
This is a project focusing on Light Measurement. By integrating Bolt IOT module, light sensor, resistor and Bolt platform , light levels in real time could be measured and monitored. One potential application of this project is to monitor the sunlight exposure for plants. By tracking and analyzing the data collected, we can ensure that our plants are receiving the necessary amount of sunlight for optimal growth.